services/main_orchestrator/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
python-multipart>=0.0.6
pillow>=10.2.0

